requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.154
botocore==1.34.154
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
json_repair==0.30.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
langchain==0.3.2
langchain-core==0.3.9
langchain-text-splitters==0.3.0
langgraph==0.2.37
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.131
logger==1.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.9.0
orjson==3.10.7
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
playwright==1.45.0
protobuf==4.25.2
pyarrow==14.0.2
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
PyPDF4==1.27.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.17.1
s3transfer==0.10.2
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.30.0
streamlit-card==1.0.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
webdriver-manager==4.0.2
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0